% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_cp.R
\name{conv_ids_cp}
\alias{conv_ids_cp}
\title{Converting gene IDS using orgDB}
\usage{
conv_ids_cp(input, from, to, db)
}
\arguments{
\item{input}{vector of the IDs to convert}

\item{from}{input ID type}

\item{to}{output ID type or vector of output ID types}

\item{db}{annotation database}
}
\value{
dataframe with the translated IDs
}
\description{
This function returns a dataframe with the translated IDs.
}
\details{
This function takes a vector of IDs and translated them from
the input format to the output format using the selected orgDB.
In case the orgDB is not installed before launching the function,
the function will exit with error.
}
\examples{
conv_ids_cp(rownames(df), "ENSEMBL", c("ENTREZID", "GENENAME"), "org.Mmu.eg.db")
conv_ids_cp(c("IFNA1", "IFNA13", "SLC2A3"), "SYMBOL", "ENSEMBL", "org.Mmu.eg.db")
}
