% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_full.R
\name{conv_ids_full}
\alias{conv_ids_full}
\title{CONVERTING IDS USING orgDB and maRt
This function returns a dataframe with translated IDs and NA row for not anootated genes.
It works only with the following IDs: SYMBOL, ENTREZID, GENENAME, ENSEMBL}
\usage{
conv_ids_full(
  input,
  db_cluster_profiler,
  from_cluster_profiler,
  mart,
  from_mart
)
}
\arguments{
\item{input}{vector of IDs}

\item{db_cluster_profiler}{annotation clusterProfiler database}

\item{from_cluster_profiler}{input clusterProfiler ID type: ENSEMBL, ENTREZID, SYMBOL}

\item{from_mart}{annotation mart}
}
\value{
dataframe with the ENSEMBL ID, SYMBOL, ENTREZID, GENENAME(description); NAs are not dropped
}
\description{
CONVERTING IDS USING orgDB and maRt
This function returns a dataframe with translated IDs and NA row for not anootated genes.
It works only with the following IDs: SYMBOL, ENTREZID, GENENAME, ENSEMBL
}
\examples{
conv_ids_full(rownames(dataNorm_df), "org.Mmu.eg.db", "mmulatta_gene_ensembl", "ENSEMBL", "ensembl_gene_id")
conv_ids_full(c("IFNA13", "SLC2A3", "CD45RA", "CDY2A", "IGHM", "IGKC"), "org.Mmu.eg.db",  "SYMBOL", "mmulatta_gene_ensembl", "external_gene_name")
}
