% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_mart.R
\name{conv_ids_mart}
\alias{conv_ids_mart}
\title{CONVERTING IDS USING maRt
This function returns a dataframe with the translated IDs}
\usage{
conv_ids_mart(input, mart, from, to)
}
\arguments{
\item{input}{vector of IDs}

\item{mart}{annotation maRt}

\item{from}{input ID type}

\item{to}{output ID type or list of output ID types}
}
\value{
deataframe with translated IDs
}
\description{
CONVERTING IDS USING maRt
This function returns a dataframe with the translated IDs
}
\examples{
conv_ids_mart(c("CDY2A","IGHM", "IGKC"), "mmulatta_gene_ensembl", "external_gene_name", c("entrezgene_id", "ensembl_gene_id"))
}
