% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_pca.R
\name{get_main_pca}
\alias{get_main_pca}
\title{Getting the main n PCA components}
\usage{
get_main_pca(df, n)
}
\arguments{
\item{df}{dataframe with normalized counts (column = sample, row = transcript)}

\item{n}{number of top contributors}
}
\value{
vector of top n contributors
}
\description{
This function returns a vector of top n contributors
}
\examples{
get_main_pca(df, 3)
}
