% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGO_all_groups.R
\name{groupGO_all_groups}
\alias{groupGO_all_groups}
\title{GO analysis and plots for BP, CC, MF}
\usage{
groupGO_all_groups(
  gene_entrezIDs,
  orgDB,
  name = "",
  bp = 2,
  cc = 3,
  mf = 3,
  xls = TRUE,
  pdf = TRUE
)
}
\arguments{
\item{gene_entrezIDs}{vector of entrezIDs}

\item{orgDB}{db to use for GO}

\item{name}{name of the output file}

\item{bp}{level of the biological process with default 2}

\item{cc}{level of the cellular component with default 3}

\item{mf}{level of the molecular function with default 3}

\item{xls}{logical parameter to create excel with with default TRUE}

\item{pdf}{logical parameter to create pdf with with default TRUE}
}
\value{
table with groupGO result arranged as list of dataframes

excel file with groupGO result and the pdf with the groupGO result barplot
per biological process, cellular component and molecular function
}
\description{
This function returns table for GO results for BP, CC, MF with gene symbols.
The function may also create excel with the GO result tables and the PDFs with the barplots of the results.
}
\examples{
groupGO_all_groups(entrez_ids_list, "org.Mmu.eg.db", "out_name", mf = 1, xls = FALSE, pdf = FALSE)
}
