% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfc_cols.R
\name{logfc_cols}
\alias{logfc_cols}
\title{Calculating logFC between two columns of a dataframe}
\usage{
logfc_cols(data, col1, col2)
}
\arguments{
\item{col1}{name or number of the column in df}

\item{col2}{name or number the column in df}

\item{df}{dataframe with normalized counts}
}
\value{
dataframe with 3 columns col1, col2, log2FC
}
\description{
This function returns dataframe with 3 columns col1, col2, log2FC.
Mind the column names order: the logFC is always calculated between the first anf the second input
}
\examples{
logfc_cols(dataNorm_df, "selected_column", "5")
logfc_cols(dataNorm_df, "selected_column", 5)
logfc_cols(dataNorm_df, 1, 5)
}
