% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{Outersect}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
sorted vector of non-overlapping values from lists x and y
}
\description{
This function returns the non-overlapping values from the two input vectors.
}
\details{
It checks for values from the first inout not appearing in the second input
and for the values from the second inout not appearing in the first input.
}
\examples{
outersect(df_1[,2],df_2[,2])
outersect(1:5, 4:7)
}
