% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_entropy.R
\name{sh_entropy}
\alias{sh_entropy}
\title{Calculating Shannon entropy of each sample, Shannnon entropy
values and transcript frequency visualization}
\usage{
sh_entropy(
  df,
  name = "Shannon entropy",
  pdf = TRUE,
  relevel = FALSE,
  levels = NULL
)
}
\arguments{
\item{df}{input dataframe with transcripts in rows and samples in columns}

\item{name}{name of the output pdf with defult "Shannon entropy", used when pdf is TRUE}

\item{pdf}{logical parameter to create pdf with with default TRUE}

\item{relevel}{logical parameter to reorder the samples order in the plots with default FALSE}

\item{levels}{vector with the correct order of the samples with default NULL; will be used only if relevel is TRUE}
}
\value{
list with the entropy levels and plots
}
\description{
This function returns a list, containing dataframe with the entropy
values per sample,
entropy plot and transcript frequency plot
}
\details{
This function takes a dataframe with transcripts in rows and samples in
columns as an input and calculates the Shannon entropy for each sample
and the frequency of the transcript in the sample. Further it visualizes
the entropy levels and the frequencies. The transcripts are reordered by
their frequencies.
Function can also print the plots to pdf if pdf parameter is TRUE.
}
\examples{
sh_entropy(df, "name", relevel = TRUE, levels = c(seq(1, ncol(df), 1)))
sh_entropy(df)
sh_entropy(df, pdf = FALSE)
}
