% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_input_type.R
\name{test_input_type}
\alias{test_input_type}
\title{CONVERTING IDS USING orgDB
This function returns a dataframe with the translated IDs}
\usage{
test_input_type(data, input)
}
\arguments{
\item{input}{string or character or variable}

\item{from}{input ID type}

\item{to}{output ID type or list of output ID types}

\item{db}{annotation database}
}
\value{
dataframe with the translated IDs
}
\description{
CONVERTING IDS USING orgDB
This function returns a dataframe with the translated IDs
}
