% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Finding gene homologs}
\usage{
translate(
  input,
  taxid,
  from_id = "ENTREZID",
  homologeneFile,
  db_cluster_profiler = NULL,
  from_mart = NULL,
  mart = NULL
)
}
\arguments{
\item{input}{vector of IDs}

\item{taxid}{taxonomy ID of the target species -- https://www.ncbi.nlm.nih.gov/Taxonomy/Browser}

\item{from_id}{input ID type with default "ENTREZID", other possible: "ENSEMBL", "SYMBOL"}

\item{homologeneFile}{homology table -- https://ftp.ncbi.nih.gov/pub/HomoloGene/current/homologene.data}

\item{db_cluster_profiler}{annotation clusterProfiler database}

\item{from_mart}{input maRt ID type: ensembl_gene_id, external_gene_name, entrezgene_id}

\item{mart}{annotation martL}
}
\value{
dataframe with the ENSEMBL ID, SYMBOL, ENTREZID, GENENAME(description); NAs are not dropped
}
\description{
This function returns a dataframe with homolog IDs and NA row for the genes with no homologs.
}
\details{
This function takes a vector of IDs. If the IDs are not ENTREZ IDs,
they are converted to the ENTREZ ID format. Next, the homologs of the IDs are found using
the homologeneFile. The IDs which were not coverted  and the IDs without homologs are returned.
}
\examples{
translate(rownames(dataNorm_df), 9606, homologeneFile = file)
translate(rownames(dataNorm_df), 9606,
    from_id = "ENSEMBL", homologeneFile = file,
    db_cluster_profiler = "org.Mmu.eg.db", mart = "mmulatta_gene_ensembl",
    from_mart = "ensembl_gene_id"
)
}
