% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_cp.R
\name{conv_ids_cp}
\alias{conv_ids_cp}
\title{CONVERTING IDS USING orgDB
This function returns a dataframe with the translated IDs}
\usage{
conv_ids_cp(input, from, to, db)
}
\arguments{
\item{input}{vector}

\item{from}{input ID type}

\item{to}{output ID type or list of output ID types}

\item{db}{annotation database}
}
\value{
dataframe with the translated IDs
}
\description{
CONVERTING IDS USING orgDB
This function returns a dataframe with the translated IDs
}
\examples{
conv_ids_cp(rownames(df), "ENSEMBL", c("ENTREZID", "GENENAME"), "org.Mmu.eg.db")
conv_ids_cp(c("IFNA1", "IFNA13", "SLC2A3"), "SYMBOL", "ENSEMBL", "org.Mmu.eg.db")
}
