% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_full.R
\name{conv_ids_full}
\alias{conv_ids_full}
\title{Converting gene IDS using orgDB and maRt}
\usage{
conv_ids_full(
  input,
  db_cluster_profiler,
  from_cluster_profiler,
  mart,
  from_mart
)
}
\arguments{
\item{input}{vector of IDs}

\item{db_cluster_profiler}{annotation clusterProfiler database}

\item{from_cluster_profiler}{input clusterProfiler ID type: ENSEMBL, ENTREZID, SYMBOL}

\item{from_mart}{annotation mart}
}
\value{
dataframe with the ENSEMBL ID, SYMBOL, ENTREZID, GENENAME(description); NAs are not dropped
}
\description{
This function returns a dataframe with translated IDs and NA row for the not anootated genes.
}
\details{
This function takes a vector of IDs and translated them from
the input format to the output format using the selected orgDB.
The IDs not found by OrgDB are further annotated by maRt.
If the ID was not found it will be at the end of the output dataframe with NAs.
This function works only with the IDs: ENSEMBL ID, SYMBOL, ENTREZID
In case the orgDB is not installed before launching the function,
the function will exit with error.
}
\examples{
conv_ids_full(rownames(dataNorm_df), "org.Mmu.eg.db", "ENSEMBL", "mmulatta_gene_ensembl", "ensembl_gene_id")
conv_ids_full(c("IFNA13", "SLC2A3", "CD45RA", "CDY2A", "IGHM", "IGKC"), "org.Mmu.eg.db",  "SYMBOL", "mmulatta_gene_ensembl", "external_gene_name")
}
