% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ids_mart.R
\name{conv_ids_mart}
\alias{conv_ids_mart}
\title{Converting gene IDS using maRt}
\usage{
conv_ids_mart(input, mart, from, to)
}
\arguments{
\item{input}{vector of IDs}

\item{mart}{annotation maRt}

\item{from}{input ID type}

\item{to}{output ID type or vector of output ID types}
}
\value{
dataframe with translated IDs
}
\description{
This function returns a dataframe with the translated IDs.
#'@details
This function takes a vector of IDs and translated them from
the input format to the output format using the selected maRt.
}
\examples{
conv_ids_mart(c("CDY2A","IGHM", "IGKC"), "mmulatta_gene_ensembl", "external_gene_name", 
c("entrezgene_id", "ensembl_gene_id"))
}
