% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{OUTERSECT OF TWO VECTORS}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
sorted vector of non-overlapping values from lists x and y
}
\description{
This function returns the non-overlapping values from two vectors
}
\examples{
outersect(df_1[,2],df_2[,2])
outersect(1:5, 4:7)
}
