% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_input_type.R
\name{test_input_type}
\alias{test_input_type}
\title{Help function for logfc_cols()}
\usage{
test_input_type(data, input)
}
\arguments{
\item{data}{input df}

\item{input}{column number or column to check}
}
\value{
error or pass
}
\description{
This function checks whether the input exists in the df:
\itemize{
\item for strings checks the column names
\item for numbers checks whether it is smaller than the number of columns
}
}
